mcp-calculator/requirements_mcp.txt
# MCP Service 依赖
fastmcp>=0.1.0
python-dotenv>=0.19.0
requests>=2.26.0
